requirements.txt
flask==2.3.2
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
faiss-cpu==1.7.4
transformers==4.31.0
sentence-transformers==2.2.2
psycopg2-binary==2.9.6
sqlalchemy==2.0.19